+++ dvc.yaml
@@ -1,1 +1,10 @@
-# we have to write here stages.
+stages:
+  load_data:
+    cmd: python src/stage_01_data_loader.py --config=config/config.yaml
+    deps:
+      - src/stage_01_data_loader.py
+      - src/utils/all_utils.py
+      - config/config.yaml
+    outs:
+      - artifacts/local_data_dirs/aps_failure_training_set.csv
+
